#68 - Moved the save encounter modal
